planet: adding end-to-end test build target
This is the initial commit that introduces end-to-end testing to the
kubernetes cluster created by planet. The tests are run inside a docker
container based on planet/buildox image which has been extended to
include the kubernetes repository in its entirety. This is not strictly
necessary and can be changed later - this is due to some tests
requiring configuration data that resides within the repo.
The tests are run using the ginkgo test runner, and the two extra
executables required for testing are also copied to build/current.