planet: e2e tests
Implemented planet test command that spawns the test runner (ginkgo)
with given flags and reports the test status back via exit code.
e2e test uses a special testbox docker image containing a copy of
kubernetes repository as it needs certain assets during runtime.
The plan is to use a smallest image possible - for instance, alpine
linux, which is ~200M (which half the size of the equivalent
debian:jessie image).