planet/e2e: implemented a self-test mode of the `start` command.
The `start` command received a `self-test` flag to run in test mode.
This way it is easy to test a new planet distribution with a single
command.
The stand-alone test (using a script to invoke the test runner) will be
recovered to support test environments other than planet (like manual
k8s vagrant configuration).