planet/e2e: re-introduced `planet test` for e2e testing independent of the planet cluster.
The planet cluster can be tested using the new self-test mode built into
`planet start`. The old `planet test` can be used to run e2e tests on a
cluster created with an environment alternative to planet.
Moved the k8s repo version to the main makefile to avoid version
mismatches when multiple files are edited and not kept in sync.