planet: renamed private-ip -> public-ip with a fitter description.
The new public-ip option specifies the IP used for other nodes to
for inter-host communication. Previously, planet used adhoc script that
collected all available interfaces into an environment file and picked
the default ipv4 interface which is not always a better choice - in
vagrant is corresponds to a eth0 bridge that cannot be used for
inter-host communication.