planet/monitoring: complemented monit service query with that of systemd.
For the time being, the systemd query relies on a simple static binary
that uses go-systemd to request the list of units and outputs them to stdout
as json - this is good enough for a start.
After both monitoring services have been queried, the results are
aggregated and also output to stdout as json.