planet: clean up stale socket file on start
Currently, if the planet is forcibly terminated - as a result of a
reboot or after a kill(1), the unix domain socket file is not cleaned up.
If the socket file has not been removed, next attempt to bind to this address
(bind(2)) will fail with EADDRINUSE (address already in use) regardless if
the socket is actually open. In this case, the file needs to be removed
before planet can successfully bind again.
Since planet already intercepts SIGTERM and SIGINT signals and shuts
down properly (in which case, with the server gone, the socket file is
also automatically removed by the library (UnixListener.Close), we only
need to handle the stale case by:
  - ensuring it is safe to remove
  - actually removing the file prior to binding to the address again
This is achieved by attempting to connect to the socket - if the connection attempt
is unsuccessful, either because the socket is not open or the file does not exist -
start up proceeds normally (and the file is removed if existent).