planet: offload socket handling to systemd if need be.
Adds support for systemd socket-activation when planet does not manage
the socket file instead relying on being passed the file descriptors to
a socket file from systemd. This can be used when planet runs as a systemd unit.
This can be further simplified if planet is only meant to be run as a
systemd unit.