planet: make service user/group configurable
This change introduces additional `start` command flags to configure
service user/group. Service user is the user used for non-privileged
services. Prior to this change, planet defaulted to user/group
`planet`/`planet` and set uid/gid to a fixed value of 1000.
This is suboptimal since it might clash with an existing user account
used for other purposes.
Additionally, there was no user/group configuration for almost all services
and as such they run as root. Now, only the services that require root privileges
are run as root - all other run under a non-privileged service user.
The user for administrative and privileged tasks is not configurable and
left implicit as root.