Work in progess updating to k8s 1.6:
Rename kube-proxy user to be the built-in system:kube-proxy since it
needs to be bound to the system:node-proxier cluster role to work
properly.

FIXME: temporarily removed kubelet's resolv-conf configuration which has
changed the implementation and now breaks Pods with in hostNetwork mode.
A proper solution is beoing sought.

Updates https://github.com/gravitational/gravity/issues/2460