Remove start dependency on docker (#433)
The registry doesn't actually require docker to be running, so remove this dependency. The latest version of docker seems to be taking longer than normal to start, which can cause a race in the installer where the registry won't be started before it's used (the installer should also be separately fixed to wait longer).