Implement workaround to clean up leaking cgroups
This change implements a cleaner, that scans for cgroups created by
systemd-run --scope that do not have any pids assigned, indicating
that the cgroup is unused and should be cleaned up. On some systems
either due to systemd or the kernel, the scope is not being cleaned
up when the pids within the scope have completed execution, leading
to an eventual memory leak.

Kubernetes uses systemd-run --scope when creating mount points,
that may require drivers to be loaded/running in a separate context
from kubelet, which allows the above leak to occur.

https://github.com/kubernetes/kubernetes/issues/70324
https://github.com/kubernetes/kubernetes/issues/64137
https://github.com/gravitational/gravity/issues/1219