lib/box: fix label selection for executables given without absolute paths.
When 'gravity exec' executed a command without specifying the absolute path to the
said command, it picked the wrong label since it attempted to match each
command directly inside the rootfs directory.
Instead, if the command is given without absolute path, match it inside
the rootfs directory by using the PATH value.